Banning duplicate search results when one is outside the time range
Multiple songs can be searched with the same ID but with different timings, the actual timing used for that ID cannot be found in this case but if one is outside the range, dont use it.

Added bannedIDs list which songs are added to when the time is outside the 15 second range. Further and previous songs using that ID are removed from the search results.